% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--structs.R
\name{as.list.cairo_matrix_t}
\alias{as.list.cairo_matrix_t}
\title{unpack cairo_matrix_t struct}
\usage{
\method{as.list}{cairo_matrix_t}(x, ...)
}
\arguments{
\item{x}{\code{cairo_matrix_t *}}

\item{...}{ignored}
}
\description{
unpack cairo_matrix_t struct
}
\seealso{
Other struct: 
\code{\link{as.list.cairo_font_extents_t}()},
\code{\link{as.list.cairo_rectangle_t}()},
\code{\link{as.list.cairo_text_extents_t}()},
\code{\link{cairo_font_extents_t}()},
\code{\link{cairo_matrix_t_update}()},
\code{\link{cairo_matrix_t}()},
\code{\link{cairo_rectangle_int_t}()},
\code{\link{cairo_rectangle_t}()},
\code{\link{cairo_text_extents_t}()}
}
\concept{struct}
