% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_content_t}
\alias{cairo_content_t}
\title{cairo_content_t}
\format{
An object of class \code{list} of length 3.
}
\usage{
cairo_content_t
}
\description{
#cairo_content_t is used to describe the content that a surface will
contain, whether color information, alpha information (translucence
vs. opacity), or both.
}
\details{
Note: The large values here are designed to keep #cairo_content_t
values distinct from #cairo_format_t values so that the
implementation can detect the error if users confuse the two types.

Since: 1.0

\describe{
\item{CAIRO_CONTENT_COLOR}{The surface will hold color content only. (Since 1.0)}
\item{CAIRO_CONTENT_ALPHA}{The surface will hold alpha content only. (Since 1.0)}
\item{CAIRO_CONTENT_COLOR_ALPHA}{The surface will hold color and alpha content. (Since 1.0)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
