% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{cairo_destroy}
\alias{cairo_destroy}
\alias{cairo_reference}
\alias{cairo_get_reference_count}
\alias{cairo_surface_destroy}
\alias{cairo_surface_reference}
\alias{cairo_surface_get_reference_count}
\alias{cairo_device_destroy}
\alias{cairo_device_reference}
\alias{cairo_device_get_reference_count}
\alias{cairo_pattern_destroy}
\alias{cairo_pattern_reference}
\alias{cairo_pattern_get_reference_count}
\alias{cairo_scaled_font_destroy}
\alias{cairo_scaled_font_reference}
\alias{cairo_scaled_font_get_reference_count}
\alias{cairo_font_face_destroy}
\alias{cairo_font_face_reference}
\alias{cairo_font_face_get_reference_count}
\alias{cairo_rectangle_list_destroy}
\alias{cairo_region_destroy}
\alias{cairo_region_reference}
\alias{cairo_font_options_destroy}
\alias{cairo_path_destroy}
\title{C-style reference counting and \code{*_destroy()} functions are not implemented in R}
\usage{
cairo_destroy(...)

cairo_reference(...)

cairo_get_reference_count(...)

cairo_surface_destroy(...)

cairo_surface_reference(...)

cairo_surface_get_reference_count(...)

cairo_device_destroy(...)

cairo_device_reference(...)

cairo_device_get_reference_count(...)

cairo_pattern_destroy(...)

cairo_pattern_reference(...)

cairo_pattern_get_reference_count(...)

cairo_scaled_font_destroy(...)

cairo_scaled_font_reference(...)

cairo_scaled_font_get_reference_count(...)

cairo_font_face_destroy(...)

cairo_font_face_reference(...)

cairo_font_face_get_reference_count(...)

cairo_rectangle_list_destroy(...)

cairo_region_destroy(...)

cairo_region_reference(...)

cairo_font_options_destroy(...)

cairo_path_destroy(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
All memory and reference counting functions are currently empty.  The R
garbage collector is used to clean-up at the end of object lifetimes
}
