% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_device_type_t}
\alias{cairo_device_type_t}
\title{cairo_device_type_t}
\format{
An object of class \code{list} of length 9.
}
\usage{
cairo_device_type_t
}
\description{
#cairo_device_type_t is used to describe the type of a given
device. The devices types are also known as "backends" within cairo.
}
\details{
The device type can be queried with cairo_device_get_type()

The various #cairo_device_t functions can be used with devices of
any type, but some backends also provide type-specific functions
that must only be called with a device of the appropriate
type. These functions have names that begin with
<literal>cairo_<emphasis>type</emphasis>_device</literal> such as
cairo_xcb_device_debug_cap_xrender_version().

The behavior of calling a type-specific function with a device of
the wrong type is undefined.

New entries may be added in future versions.

Since: 1.10

\describe{
\item{CAIRO_DEVICE_TYPE_DRM}{The device is of type Direct Render Manager, since 1.10}
\item{CAIRO_DEVICE_TYPE_GL}{The device is of type OpenGL, since 1.10}
\item{CAIRO_DEVICE_TYPE_SCRIPT}{The device is of type script, since 1.10}
\item{CAIRO_DEVICE_TYPE_XCB}{The device is of type xcb, since 1.10}
\item{CAIRO_DEVICE_TYPE_XLIB}{The device is of type xlib, since 1.10}
\item{CAIRO_DEVICE_TYPE_XML}{The device is of type XML, since 1.10}
\item{CAIRO_DEVICE_TYPE_COGL}{The device is of type cogl, since 1.12}
\item{CAIRO_DEVICE_TYPE_WIN32}{The device is of type win32, since 1.12}
\item{CAIRO_DEVICE_TYPE_INVALID}{The device is invalid, since 1.10}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
