% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_extend_t}
\alias{cairo_extend_t}
\title{cairo_extend_t}
\format{
An object of class \code{list} of length 4.
}
\usage{
cairo_extend_t
}
\description{
#cairo_extend_t is used to describe how pattern color/alpha will be
determined for areas "outside" the pattern's natural area, (for
example, outside the surface bounds or outside the gradient
geometry).
}
\details{
Mesh patterns are not affected by the extend mode.

The default extend mode is \%CAIRO_EXTEND_NONE for surface patterns
and \%CAIRO_EXTEND_PAD for gradient patterns.

New entries may be added in future versions.

Since: 1.0

\describe{
\item{CAIRO_EXTEND_NONE}{pixels outside of the source pattern are fully transparent (Since 1.0)}
\item{CAIRO_EXTEND_REPEAT}{the pattern is tiled by repeating (Since 1.0)}
\item{CAIRO_EXTEND_REFLECT}{the pattern is tiled by reflecting at the edges (Since 1.0; but only implemented for surface patterns since 1.6)}
\item{CAIRO_EXTEND_PAD}{pixels outside of the pattern copy the closest pixel from the source (Since 1.2; but only implemented for surface patterns since 1.6)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
