% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--structs.R
\name{cairo_font_extents_t}
\alias{cairo_font_extents_t}
\title{Create an empty cairo_text_extents_t pointer}
\usage{
cairo_font_extents_t()
}
\description{
\code{cairo_text_extents_t * cairo_text_extents_t(void)}
}
\seealso{
Other struct: 
\code{\link{as.list.cairo_font_extents_t}()},
\code{\link{as.list.cairo_matrix_t}()},
\code{\link{as.list.cairo_rectangle_t}()},
\code{\link{as.list.cairo_text_extents_t}()},
\code{\link{cairo_matrix_t_update}()},
\code{\link{cairo_matrix_t}()},
\code{\link{cairo_rectangle_int_t}()},
\code{\link{cairo_rectangle_t}()},
\code{\link{cairo_text_extents_t}()}
}
\concept{struct}
