% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-face.R
\name{cairo_font_face_get_type}
\alias{cairo_font_face_get_type}
\title{cairo_font_face_get_type}
\usage{
cairo_font_face_get_type(font_face)
}
\arguments{
\item{font_face}{[\code{cairo_font_face_t *}] a font face}
}
\value{
[\code{int}]
}
\description{
This function returns the type of the backend used to create
a font face. See #cairo_font_type_t for available types.
}
\details{
Return value: The type of font_face.

Since: 1.2

C function prototype: \code{cairo_font_type_t cairo_font_face_get_type (cairo_font_face_t *font_face)}
}
\seealso{
Other cairo-font-face: 
\code{\link{cairo_font_face_status}()}
}
\concept{cairo-font-face}
