% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-face.R
\name{cairo_font_face_status}
\alias{cairo_font_face_status}
\title{cairo_font_face_status}
\usage{
cairo_font_face_status(font_face)
}
\arguments{
\item{font_face}{[\code{cairo_font_face_t *}] a #cairo_font_face_t}
}
\value{
[\code{int}]
}
\description{
Checks whether an error has previously occurred for this
font face
}
\details{
Return value: \%CAIRO_STATUS_SUCCESS or another error such as
\%CAIRO_STATUS_NO_MEMORY.

Since: 1.0

C function prototype: \code{cairo_status_t cairo_font_face_status (cairo_font_face_t *font_face)}
}
\seealso{
Other cairo-font-face: 
\code{\link{cairo_font_face_get_type}()}
}
\concept{cairo-font-face}
