% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_copy}
\alias{cairo_font_options_copy}
\title{cairo_font_options_copy}
\usage{
cairo_font_options_copy(original)
}
\arguments{
\item{original}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}
}
\value{
[\code{cairo_font_options_t *}]
}
\description{
Allocates a new font options object copying the option values from
original.
}
\details{
Return value: a newly allocated #cairo_font_options_t. Free with
cairo_font_options_destroy(). This function always returns a
valid pointer; if memory cannot be allocated, then a special
error object is returned where all operations on the object do nothing.
You can check for this with cairo_font_options_status().

Since: 1.0

C function prototype: \code{cairo_font_options_t * cairo_font_options_copy (const cairo_font_options_t *original)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
