% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_get_hint_style}
\alias{cairo_font_options_get_hint_style}
\title{cairo_font_options_get_hint_style}
\usage{
cairo_font_options_get_hint_style(options)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}
}
\value{
[\code{int}]
}
\description{
Gets the hint style for font outlines for the font options object.
See the documentation for #cairo_hint_style_t for full details.
}
\details{
Return value: the hint style for the font options object

Since: 1.0

C function prototype: \code{cairo_hint_style_t cairo_font_options_get_hint_style (const cairo_font_options_t *options)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
