% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_get_variations}
\alias{cairo_font_options_get_variations}
\title{cairo_font_options_get_variations}
\usage{
cairo_font_options_get_variations(options)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}
}
\value{
[\code{const char *}]
}
\description{
Gets the OpenType font variations for the font options object.
See cairo_font_options_set_variations() for details about the
string format.
}
\details{
Return value: the font variations for the font options object. The
returned string belongs to the options and must not be modified.
It is valid until either the font options object is destroyed or
the font variations in this object is modified with
cairo_font_options_set_variations().

Since: 1.16

C function prototype: \code{const char * cairo_font_options_get_variations (cairo_font_options_t *options)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
