% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_hash}
\alias{cairo_font_options_hash}
\title{cairo_font_options_hash}
\usage{
cairo_font_options_hash(options)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}
}
\value{
[\code{unsigned long}]
}
\description{
Compute a hash for the font options object; this value will
be useful when storing an object containing a #cairo_font_options_t
in a hash table.
}
\details{
Return value: the hash value for the font options object.
The return value can be cast to a 32-bit type if a
32-bit hash value is needed.

Since: 1.0

C function prototype: \code{unsigned long cairo_font_options_hash (const cairo_font_options_t *options)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
