% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_merge}
\alias{cairo_font_options_merge}
\title{cairo_font_options_merge}
\usage{
cairo_font_options_merge(options, other)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}

\item{other}{[\code{cairo_font_options_t *}] another #cairo_font_options_t}
}
\description{
Merges non-default options from other into options, replacing
existing values. This operation can be thought of as somewhat
similar to compositing other onto options with the operation
of \%CAIRO_OPERATOR_OVER.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_font_options_merge (cairo_font_options_t *options, const cairo_font_options_t *other)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
