% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_set_hint_style}
\alias{cairo_font_options_set_hint_style}
\title{cairo_font_options_set_hint_style}
\usage{
cairo_font_options_set_hint_style(options, hint_style)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}

\item{hint_style}{[\code{int}] the new hint style}
}
\description{
Sets the hint style for font outlines for the font options object.
This controls whether to fit font outlines to the pixel grid,
and if so, whether to optimize for fidelity or contrast.
See the documentation for #cairo_hint_style_t for full details.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_font_options_set_hint_style (cairo_font_options_t *options, cairo_hint_style_t hint_style)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
