% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_set_subpixel_order}
\alias{cairo_font_options_set_subpixel_order}
\title{cairo_font_options_set_subpixel_order}
\usage{
cairo_font_options_set_subpixel_order(options, subpixel_order)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}

\item{subpixel_order}{[\code{int}] the new subpixel order}
}
\description{
Sets the subpixel order for the font options object. The subpixel
order specifies the order of color elements within each pixel on
the display device when rendering with an antialiasing mode of
\%CAIRO_ANTIALIAS_SUBPIXEL. See the documentation for
#cairo_subpixel_order_t for full details.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_font_options_set_subpixel_order (cairo_font_options_t *options, cairo_subpixel_order_t subpixel_order)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_variations}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
