% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-font-options.R
\name{cairo_font_options_set_variations}
\alias{cairo_font_options_set_variations}
\title{cairo_font_options_set_variations}
\usage{
cairo_font_options_set_variations(options, variations)
}
\arguments{
\item{options}{[\code{cairo_font_options_t *}] a #cairo_font_options_t}

\item{variations}{[\code{char *}] the new font variations, or \%NULL}
}
\description{
Sets the OpenType font variations for the font options object.
Font variations are specified as a string with a format that
is similar to the CSS font-variation-settings. The string contains
a comma-separated list of axis assignments, which each assignment
consists of a 4-character axis name and a value, separated by
whitespace and optional equals sign.
}
\details{
Examples:

wght=200,wdth=140.5

wght 200 , wdth 140.5

Since: 1.16

C function prototype: \code{void cairo_font_options_set_variations (cairo_font_options_t *options, const char *variations)}
}
\seealso{
Other cairo-font-options: 
\code{\link{cairo_font_options_copy}()},
\code{\link{cairo_font_options_create}()},
\code{\link{cairo_font_options_equal}()},
\code{\link{cairo_font_options_get_antialias}()},
\code{\link{cairo_font_options_get_hint_metrics}()},
\code{\link{cairo_font_options_get_hint_style}()},
\code{\link{cairo_font_options_get_subpixel_order}()},
\code{\link{cairo_font_options_get_variations}()},
\code{\link{cairo_font_options_hash}()},
\code{\link{cairo_font_options_merge}()},
\code{\link{cairo_font_options_set_antialias}()},
\code{\link{cairo_font_options_set_hint_metrics}()},
\code{\link{cairo_font_options_set_hint_style}()},
\code{\link{cairo_font_options_set_subpixel_order}()},
\code{\link{cairo_font_options_status}()}
}
\concept{cairo-font-options}
