% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface.R
\name{cairo_format_stride_for_width}
\alias{cairo_format_stride_for_width}
\title{cairo_format_stride_for_width}
\usage{
cairo_format_stride_for_width(format, width)
}
\arguments{
\item{format}{[\code{int}] A #cairo_format_t value}

\item{width}{[\code{int}] The desired width of an image surface to be created.}
}
\value{
[\code{int}]
}
\description{
This function provides a stride value that will respect all
alignment requirements of the accelerated image-rendering code
within cairo. Typical usage will be of the form:
}
\details{
<informalexample><programlisting>
int stride;
unsigned char *data;
cairo_surface_t *surface;

stride = cairo_format_stride_for_width (format, width);
data = malloc (stride * height);
surface = cairo_image_surface_create_for_data (data, format,
width, height,
stride);
</programlisting></informalexample>

Return value: the appropriate stride to use given the desired
format and width, or -1 if either the format is invalid or the width
too large.

Since: 1.6

C function prototype: \code{int cairo_format_stride_for_width (cairo_format_t format, int width)}
}
\seealso{
Other cairo-image-surface: 
\code{\link{cairo_image_surface_create_for_data}()},
\code{\link{cairo_image_surface_create}()},
\code{\link{cairo_image_surface_get_data}()},
\code{\link{cairo_image_surface_get_format}()},
\code{\link{cairo_image_surface_get_height}()},
\code{\link{cairo_image_surface_get_stride}()},
\code{\link{cairo_image_surface_get_width}()}
}
\concept{cairo-image-surface}
