% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_format_t}
\alias{cairo_format_t}
\title{cairo_format_t}
\format{
An object of class \code{list} of length 9.
}
\usage{
cairo_format_t
}
\description{
#cairo_format_t is used to identify the memory format of
image data.
}
\details{
New entries may be added in future versions.

Since: 1.0

\describe{
\item{CAIRO_FORMAT_INVALID}{no such format exists or is supported.}
\item{CAIRO_FORMAT_ARGB32}{each pixel is a 32-bit quantity, with alpha in the upper 8 bits, then red, then green, then blue. The 32-bit quantities are stored native-endian. Pre-multiplied alpha is used. (That is, 50\% transparent red is 0x80800000, not 0x80ff0000.) (Since 1.0)}
\item{CAIRO_FORMAT_RGB24}{each pixel is a 32-bit quantity, with the upper 8 bits unused. Red, Green, and Blue are stored in the remaining 24 bits in that order. (Since 1.0)}
\item{CAIRO_FORMAT_A8}{each pixel is a 8-bit quantity holding an alpha value. (Since 1.0)}
\item{CAIRO_FORMAT_A1}{each pixel is a 1-bit quantity holding an alpha value. Pixels are packed together into 32-bit quantities. The ordering of the bits matches the endianness of the platform. On a big-endian machine, the first pixel is in the uppermost bit, on a little-endian machine the first pixel is in the least-significant bit. (Since 1.0)}
\item{CAIRO_FORMAT_RGB16_565}{each pixel is a 16-bit quantity with red in the upper 5 bits, then green in the middle 6 bits, and blue in the lower 5 bits. (Since 1.2)}
\item{CAIRO_FORMAT_RGB30}{like RGB24 but with 10bpc. (Since 1.12)}
\item{CAIRO_FORMAT_RGB96F}{3 floats, R, G, B. (Since 1.17.2)}
\item{CAIRO_FORMAT_RGBA128F}{4 floats, R, G, B, A. (Since 1.17.2)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
