% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_hint_style_t}
\alias{cairo_hint_style_t}
\title{cairo_hint_style_t}
\format{
An object of class \code{list} of length 5.
}
\usage{
cairo_hint_style_t
}
\description{
Specifies the type of hinting to do on font outlines. Hinting
is the process of fitting outlines to the pixel grid in order
to improve the appearance of the result. Since hinting outlines
involves distorting them, it also reduces the faithfulness
to the original outline shapes. Not all of the outline hinting
styles are supported by all font backends.
}
\details{
New entries may be added in future versions.

Since: 1.0

\describe{
\item{CAIRO_HINT_STYLE_DEFAULT}{Use the default hint style for font backend and target device, since 1.0}
\item{CAIRO_HINT_STYLE_NONE}{Do not hint outlines, since 1.0}
\item{CAIRO_HINT_STYLE_SLIGHT}{Hint outlines slightly to improve contrast while retaining good fidelity to the original shapes, since 1.0}
\item{CAIRO_HINT_STYLE_MEDIUM}{Hint outlines with medium strength giving a compromise between fidelity to the original shapes and contrast, since 1.0}
\item{CAIRO_HINT_STYLE_FULL}{Hint outlines to maximize contrast, since 1.0}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
