% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-png.R
\name{cairo_image_surface_create_from_png}
\alias{cairo_image_surface_create_from_png}
\title{cairo_image_surface_create_from_png}
\usage{
cairo_image_surface_create_from_png(filename)
}
\arguments{
\item{filename}{[\code{char *}] name of PNG file to load. On Windows this filename is encoded in UTF-8.}
}
\value{
[\code{cairo_surface_t *}]
}
\description{
Creates a new image surface and initializes the contents to the
given PNG file.
}
\details{
Return value: a new #cairo_surface_t initialized with the contents
of the PNG file, or a "nil" surface if any error occurred. A nil
surface can be checked for with cairo_surface_status(surface) which
may return one of the following values:

\%CAIRO_STATUS_NO_MEMORY
\%CAIRO_STATUS_FILE_NOT_FOUND
\%CAIRO_STATUS_READ_ERROR
\%CAIRO_STATUS_PNG_ERROR

Alternatively, you can allow errors to propagate through the drawing
operations and check the status on the context upon completion
using cairo_status().

Since: 1.0

C function prototype: \code{cairo_surface_t * cairo_image_surface_create_from_png (const char *filename)}
}
\seealso{
Other cairo-png: 
\code{\link{cairo_surface_write_to_png}()}
}
\concept{cairo-png}
