% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface-extra.R
\name{cairo_image_surface_get_array}
\alias{cairo_image_surface_get_array}
\title{Get a copy of the given surface as an array}
\usage{
cairo_image_surface_get_array(surface, nchannel = 4)
}
\arguments{
\item{surface}{\code{cairo_surface_t *}}

\item{nchannel}{1 = grey, 3 = rgb, 4 = rgba}
}
\value{
matrix or array of real numbers in range [0, 1]
}
\description{
\code{ unsigned char * cairo_image_surface_get_data (cairo_surface_t *surface) }
}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_polygon}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}
}
\concept{bespoke}
