% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface.R
\name{cairo_image_surface_get_format}
\alias{cairo_image_surface_get_format}
\title{cairo_image_surface_get_format}
\usage{
cairo_image_surface_get_format(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a #cairo_image_surface_t}
}
\value{
[\code{int}]
}
\description{
Get the format of the surface.
}
\details{
Return value: the format of the surface

Since: 1.2

C function prototype: \code{cairo_format_t cairo_image_surface_get_format (cairo_surface_t *surface)}
}
\seealso{
Other cairo-image-surface: 
\code{\link{cairo_format_stride_for_width}()},
\code{\link{cairo_image_surface_create_for_data}()},
\code{\link{cairo_image_surface_create}()},
\code{\link{cairo_image_surface_get_data}()},
\code{\link{cairo_image_surface_get_height}()},
\code{\link{cairo_image_surface_get_stride}()},
\code{\link{cairo_image_surface_get_width}()}
}
\concept{cairo-image-surface}
