% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface.R
\name{cairo_image_surface_get_height}
\alias{cairo_image_surface_get_height}
\title{cairo_image_surface_get_height}
\usage{
cairo_image_surface_get_height(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a #cairo_image_surface_t}
}
\value{
[\code{int}]
}
\description{
Get the height of the image surface in pixels.
}
\details{
Return value: the height of the surface in pixels.

Since: 1.0

C function prototype: \code{int cairo_image_surface_get_height (cairo_surface_t *surface)}
}
\seealso{
Other cairo-image-surface: 
\code{\link{cairo_format_stride_for_width}()},
\code{\link{cairo_image_surface_create_for_data}()},
\code{\link{cairo_image_surface_create}()},
\code{\link{cairo_image_surface_get_data}()},
\code{\link{cairo_image_surface_get_format}()},
\code{\link{cairo_image_surface_get_stride}()},
\code{\link{cairo_image_surface_get_width}()}
}
\concept{cairo-image-surface}
