% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface-extra.R
\name{cairo_image_surface_get_raw}
\alias{cairo_image_surface_get_raw}
\title{Get a \emph{copy} of the given surface as packed color in a vector of raw values}
\usage{
cairo_image_surface_get_raw(surface)
}
\arguments{
\item{surface}{\code{cairo_surface_t *}}
}
\value{
matrix or array of real numbers in range [0, 1]
}
\description{
Get a \emph{copy} of the given surface as packed color in a vector of raw values
}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_polygon}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}
}
\concept{bespoke}
