% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface.R
\name{cairo_image_surface_get_stride}
\alias{cairo_image_surface_get_stride}
\title{cairo_image_surface_get_stride}
\usage{
cairo_image_surface_get_stride(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a #cairo_image_surface_t}
}
\value{
[\code{int}]
}
\description{
Get the stride of the image surface in bytes
}
\details{
Return value: the stride of the image surface in bytes (or 0 if
surface is not an image surface). The stride is the distance in
bytes from the beginning of one row of the image data to the
beginning of the next row.

Since: 1.2

C function prototype: \code{int cairo_image_surface_get_stride (cairo_surface_t *surface)}
}
\seealso{
Other cairo-image-surface: 
\code{\link{cairo_format_stride_for_width}()},
\code{\link{cairo_image_surface_create_for_data}()},
\code{\link{cairo_image_surface_create}()},
\code{\link{cairo_image_surface_get_data}()},
\code{\link{cairo_image_surface_get_format}()},
\code{\link{cairo_image_surface_get_height}()},
\code{\link{cairo_image_surface_get_width}()}
}
\concept{cairo-image-surface}
