% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-image-surface-extra.R
\name{cairo_image_surface_write_array}
\alias{cairo_image_surface_write_array}
\title{Write an array onto an existing surface}
\usage{
cairo_image_surface_write_array(surface, arr, maxval = 1)
}
\arguments{
\item{surface}{cairo surface}

\item{arr}{matrix, RGB array or RGBA array. numeric/double/real values only}

\item{maxval}{maximum value of colour representation. Usually 1.}
}
\description{
\code{ int cairo_image_surface_write_robj(cairo_surface_t * surface, unsigned char *data, int width, int height) }
}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_polygon}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}
}
\concept{bespoke}
