% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_line_join_t}
\alias{cairo_line_join_t}
\title{cairo_line_join_t}
\format{
An object of class \code{list} of length 3.
}
\usage{
cairo_line_join_t
}
\description{
Specifies how to render the junction of two lines when stroking.
}
\details{
The default line join style is \%CAIRO_LINE_JOIN_MITER.

Since: 1.0

\describe{
\item{CAIRO_LINE_JOIN_MITER}{use a sharp (angled) corner, see cairo_set_miter_limit() (Since 1.0)}
\item{CAIRO_LINE_JOIN_ROUND}{use a rounded join, the center of the circle is the joint point (Since 1.0)}
\item{CAIRO_LINE_JOIN_BEVEL}{use a cut-off join, the join is cut off at half the line width from the joint point (Since 1.0)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
