% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-matrix.R
\name{cairo_matrix_init}
\alias{cairo_matrix_init}
\title{cairo_matrix_init}
\usage{
cairo_matrix_init(matrix, xx, yx, xy, yy, x0, y0)
}
\arguments{
\item{matrix}{[\code{cairo_matrix_t *}] a #cairo_matrix_t}

\item{xx}{[\code{double}] xx component of the affine transformation}

\item{yx}{[\code{double}] yx component of the affine transformation}

\item{xy}{[\code{double}] xy component of the affine transformation}

\item{yy}{[\code{double}] yy component of the affine transformation}

\item{x0}{[\code{double}] X translation component of the affine transformation}

\item{y0}{[\code{double}] Y translation component of the affine transformation}
}
\description{
Sets matrix to be the affine transformation given by
xx, yx, xy, yy, x0, y0. The transformation is given
by:
<programlisting>
x_new = xx * x + xy * y + x0;
y_new = yx * x + yy * y + y0;
</programlisting>
}
\details{
Since: 1.0

C function prototype: \code{void cairo_matrix_init (cairo_matrix_t *matrix, double xx, double yx, double xy, double yy, double x0, double y0)}
}
\seealso{
Other cairo-matrix: 
\code{\link{cairo_matrix_init_rotate}()},
\code{\link{cairo_matrix_init_scale}()},
\code{\link{cairo_matrix_init_translate}()},
\code{\link{cairo_matrix_invert}()},
\code{\link{cairo_matrix_multiply}()},
\code{\link{cairo_matrix_rotate}()},
\code{\link{cairo_matrix_scale}()},
\code{\link{cairo_matrix_transform_distance}()},
\code{\link{cairo_matrix_transform_point}()},
\code{\link{cairo_matrix_translate}()}
}
\concept{cairo-matrix}
