% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-matrix.R
\name{cairo_matrix_init_scale}
\alias{cairo_matrix_init_scale}
\title{cairo_matrix_init_scale}
\usage{
cairo_matrix_init_scale(matrix, sx, sy)
}
\arguments{
\item{matrix}{[\code{cairo_matrix_t *}] a #cairo_matrix_t}

\item{sx}{[\code{double}] scale factor in the X direction}

\item{sy}{[\code{double}] scale factor in the Y direction}
}
\description{
Initializes matrix to a transformation that scales by sx and sy
in the X and Y dimensions, respectively.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_matrix_init_scale (cairo_matrix_t *matrix, double sx, double sy)}
}
\seealso{
Other cairo-matrix: 
\code{\link{cairo_matrix_init_rotate}()},
\code{\link{cairo_matrix_init_translate}()},
\code{\link{cairo_matrix_init}()},
\code{\link{cairo_matrix_invert}()},
\code{\link{cairo_matrix_multiply}()},
\code{\link{cairo_matrix_rotate}()},
\code{\link{cairo_matrix_scale}()},
\code{\link{cairo_matrix_transform_distance}()},
\code{\link{cairo_matrix_transform_point}()},
\code{\link{cairo_matrix_translate}()}
}
\concept{cairo-matrix}
