% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-matrix.R
\name{cairo_matrix_invert}
\alias{cairo_matrix_invert}
\title{cairo_matrix_invert}
\usage{
cairo_matrix_invert(matrix)
}
\arguments{
\item{matrix}{[\code{cairo_matrix_t *}] a #cairo_matrix_t}
}
\value{
[\code{int}]
}
\description{
Changes matrix to be the inverse of its original value. Not
all transformation matrices have inverses; if the matrix
collapses points together (it is <firstterm>degenerate</firstterm>),
then it has no inverse and this function will fail.
}
\details{
Returns: If matrix has an inverse, modifies matrix to
be the inverse matrix and returns \%CAIRO_STATUS_SUCCESS. Otherwise,
returns \%CAIRO_STATUS_INVALID_MATRIX.

Since: 1.0

C function prototype: \code{cairo_status_t cairo_matrix_invert (cairo_matrix_t *matrix)}
}
\seealso{
Other cairo-matrix: 
\code{\link{cairo_matrix_init_rotate}()},
\code{\link{cairo_matrix_init_scale}()},
\code{\link{cairo_matrix_init_translate}()},
\code{\link{cairo_matrix_init}()},
\code{\link{cairo_matrix_multiply}()},
\code{\link{cairo_matrix_rotate}()},
\code{\link{cairo_matrix_scale}()},
\code{\link{cairo_matrix_transform_distance}()},
\code{\link{cairo_matrix_transform_point}()},
\code{\link{cairo_matrix_translate}()}
}
\concept{cairo-matrix}
