% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--structs.R
\name{cairo_matrix_t}
\alias{cairo_matrix_t}
\title{cairo_matrix_t}
\usage{
cairo_matrix_t(xx, yx, xy, yy, x0, y0)
}
\arguments{
\item{xx}{xx component of the affine transformation}

\item{yx}{yx component of the affine transformation}

\item{xy}{xy component of the affine transformation}

\item{yy}{yy component of the affine transformation}

\item{x0}{X translation component of the affine transformation}

\item{y0}{Y translation component of the affine transformation}
}
\description{
A #cairo_matrix_t holds an affine transformation, such as a scale,
rotation, shear, or a combination of those. The transformation of
a point (x, y) is given by:
<programlisting>
x_new = xx * x + xy * y + x0;
y_new = yx * x + yy * y + y0;
</programlisting>
}
\details{
Since: 1.0
}
\seealso{
Other struct: 
\code{\link{as.list.cairo_font_extents_t}()},
\code{\link{as.list.cairo_matrix_t}()},
\code{\link{as.list.cairo_rectangle_t}()},
\code{\link{as.list.cairo_text_extents_t}()},
\code{\link{cairo_font_extents_t}()},
\code{\link{cairo_matrix_t_update}()},
\code{\link{cairo_rectangle_int_t}()},
\code{\link{cairo_rectangle_t}()},
\code{\link{cairo_text_extents_t}()}
}
\concept{struct}
