% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--structs.R
\name{cairo_matrix_t_update}
\alias{cairo_matrix_t_update}
\title{Update the values in a cairo_matrix_t}
\usage{
cairo_matrix_t_update(
  mat,
  xx = NULL,
  yx = NULL,
  xy = NULL,
  yy = NULL,
  x0 = NULL,
  y0 = NULL
)
}
\arguments{
\item{mat}{cairo_matrix_t object to update}

\item{xx}{xx component of the affine transformation}

\item{yx}{yx component of the affine transformation}

\item{xy}{xy component of the affine transformation}

\item{yy}{yy component of the affine transformation}

\item{x0}{X translation component of the affine transformation}

\item{y0}{Y translation component of the affine transformation}
}
\description{
Update the values in a cairo_matrix_t
}
\seealso{
Other struct: 
\code{\link{as.list.cairo_font_extents_t}()},
\code{\link{as.list.cairo_matrix_t}()},
\code{\link{as.list.cairo_rectangle_t}()},
\code{\link{as.list.cairo_text_extents_t}()},
\code{\link{cairo_font_extents_t}()},
\code{\link{cairo_matrix_t}()},
\code{\link{cairo_rectangle_int_t}()},
\code{\link{cairo_rectangle_t}()},
\code{\link{cairo_text_extents_t}()}
}
\concept{struct}
