% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-matrix.R
\name{cairo_matrix_transform_point}
\alias{cairo_matrix_transform_point}
\title{cairo_matrix_transform_point}
\usage{
cairo_matrix_transform_point(matrix, x, y)
}
\arguments{
\item{matrix}{[\code{cairo_matrix_t *}] a #cairo_matrix_t}

\item{x}{[\code{double *}] X position. An in/out parameter}

\item{y}{[\code{double *}] Y position. An in/out parameter}
}
\description{
Transforms the point (x, y) by matrix.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_matrix_transform_point (const cairo_matrix_t *matrix, double *x, double *y)}
}
\seealso{
Other cairo-matrix: 
\code{\link{cairo_matrix_init_rotate}()},
\code{\link{cairo_matrix_init_scale}()},
\code{\link{cairo_matrix_init_translate}()},
\code{\link{cairo_matrix_init}()},
\code{\link{cairo_matrix_invert}()},
\code{\link{cairo_matrix_multiply}()},
\code{\link{cairo_matrix_rotate}()},
\code{\link{cairo_matrix_scale}()},
\code{\link{cairo_matrix_transform_distance}()},
\code{\link{cairo_matrix_translate}()}
}
\concept{cairo-matrix}
