% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-matrix.R
\name{cairo_matrix_translate}
\alias{cairo_matrix_translate}
\title{cairo_matrix_translate}
\usage{
cairo_matrix_translate(matrix, tx, ty)
}
\arguments{
\item{matrix}{[\code{cairo_matrix_t *}] a #cairo_matrix_t}

\item{tx}{[\code{double}] amount to translate in the X direction}

\item{ty}{[\code{double}] amount to translate in the Y direction}
}
\description{
Applies a translation by tx, ty to the transformation in
matrix. The effect of the new transformation is to first translate
the coordinates by tx and ty, then apply the original transformation
to the coordinates.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_matrix_translate (cairo_matrix_t *matrix, double tx, double ty)}
}
\seealso{
Other cairo-matrix: 
\code{\link{cairo_matrix_init_rotate}()},
\code{\link{cairo_matrix_init_scale}()},
\code{\link{cairo_matrix_init_translate}()},
\code{\link{cairo_matrix_init}()},
\code{\link{cairo_matrix_invert}()},
\code{\link{cairo_matrix_multiply}()},
\code{\link{cairo_matrix_rotate}()},
\code{\link{cairo_matrix_scale}()},
\code{\link{cairo_matrix_transform_distance}()},
\code{\link{cairo_matrix_transform_point}()}
}
\concept{cairo-matrix}
