% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cairo_mesh_pattern_begin_patch}
\alias{cairo_mesh_pattern_begin_patch}
\title{cairo_mesh_pattern_begin_patch}
\usage{
cairo_mesh_pattern_begin_patch(pattern)
}
\arguments{
\item{pattern}{[\code{cairo_pattern_t *}] a #cairo_pattern_t}
}
\description{
Begin a patch in a mesh pattern.
}
\details{
After calling this function, the patch shape should be defined with
cairo_mesh_pattern_move_to(), cairo_mesh_pattern_line_to() and
cairo_mesh_pattern_curve_to().

After defining the patch, cairo_mesh_pattern_end_patch() must be
called before using pattern as a source or mask.

Note: If pattern is not a mesh pattern then pattern will be put
into an error status with a status of
\%CAIRO_STATUS_PATTERN_TYPE_MISMATCH. If pattern already has a
current patch, it will be put into an error status with a status of
\%CAIRO_STATUS_INVALID_MESH_CONSTRUCTION.

Since: 1.12

C function prototype: \code{void cairo_mesh_pattern_begin_patch (cairo_pattern_t *pattern)}
}
\seealso{
Other cairo-pattern: 
\code{\link{cairo_mesh_pattern_curve_to}()},
\code{\link{cairo_mesh_pattern_end_patch}()},
\code{\link{cairo_mesh_pattern_get_control_point}()},
\code{\link{cairo_mesh_pattern_get_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_get_patch_count}()},
\code{\link{cairo_mesh_pattern_get_path}()},
\code{\link{cairo_mesh_pattern_line_to}()},
\code{\link{cairo_mesh_pattern_move_to}()},
\code{\link{cairo_mesh_pattern_set_control_point}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgb}()},
\code{\link{cairo_pattern_add_color_stop_rgba}()},
\code{\link{cairo_pattern_add_color_stop_rgb}()},
\code{\link{cairo_pattern_create_for_surface}()},
\code{\link{cairo_pattern_create_linear}()},
\code{\link{cairo_pattern_create_mesh}()},
\code{\link{cairo_pattern_create_radial}()},
\code{\link{cairo_pattern_create_rgba}()},
\code{\link{cairo_pattern_create_rgb}()},
\code{\link{cairo_pattern_get_color_stop_count}()},
\code{\link{cairo_pattern_get_color_stop_rgba}()},
\code{\link{cairo_pattern_get_extend}()},
\code{\link{cairo_pattern_get_filter}()},
\code{\link{cairo_pattern_get_linear_points}()},
\code{\link{cairo_pattern_get_matrix}()},
\code{\link{cairo_pattern_get_radial_circles}()},
\code{\link{cairo_pattern_get_rgba}()},
\code{\link{cairo_pattern_get_type}()},
\code{\link{cairo_pattern_set_extend}()},
\code{\link{cairo_pattern_set_filter}()},
\code{\link{cairo_pattern_set_matrix}()},
\code{\link{cairo_pattern_status}()}
}
\concept{cairo-pattern}
