% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cairo_pattern_get_color_stop_rgba}
\alias{cairo_pattern_get_color_stop_rgba}
\title{cairo_pattern_get_color_stop_rgba}
\usage{
cairo_pattern_get_color_stop_rgba(
  pattern,
  index,
  offset,
  red,
  green,
  blue,
  alpha
)
}
\arguments{
\item{pattern}{[\code{cairo_pattern_t *}] a #cairo_pattern_t}

\item{index}{[\code{int}] index of the stop to return data for}

\item{offset}{[\code{double *}] return value for the offset of the stop, or \%NULL}

\item{red}{[\code{double *}] return value for red component of color, or \%NULL}

\item{green}{[\code{double *}] return value for green component of color, or \%NULL}

\item{blue}{[\code{double *}] return value for blue component of color, or \%NULL}

\item{alpha}{[\code{double *}] return value for alpha component of color, or \%NULL}
}
\value{
[\code{int}]
}
\description{
Gets the color and offset information at the given index for a
gradient pattern.  Values of index range from 0 to n-1
where n is the number returned
by cairo_pattern_get_color_stop_count().
}
\details{
Return value: \%CAIRO_STATUS_SUCCESS, or \%CAIRO_STATUS_INVALID_INDEX
if index is not valid for the given pattern.  If the pattern is
not a gradient pattern, \%CAIRO_STATUS_PATTERN_TYPE_MISMATCH is
returned.

Since: 1.4

C function prototype: \code{cairo_status_t cairo_pattern_get_color_stop_rgba (cairo_pattern_t *pattern, int index, double *offset, double *red, double *green, double *blue, double *alpha)}
}
\seealso{
Other cairo-pattern: 
\code{\link{cairo_mesh_pattern_begin_patch}()},
\code{\link{cairo_mesh_pattern_curve_to}()},
\code{\link{cairo_mesh_pattern_end_patch}()},
\code{\link{cairo_mesh_pattern_get_control_point}()},
\code{\link{cairo_mesh_pattern_get_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_get_patch_count}()},
\code{\link{cairo_mesh_pattern_get_path}()},
\code{\link{cairo_mesh_pattern_line_to}()},
\code{\link{cairo_mesh_pattern_move_to}()},
\code{\link{cairo_mesh_pattern_set_control_point}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgb}()},
\code{\link{cairo_pattern_add_color_stop_rgba}()},
\code{\link{cairo_pattern_add_color_stop_rgb}()},
\code{\link{cairo_pattern_create_for_surface}()},
\code{\link{cairo_pattern_create_linear}()},
\code{\link{cairo_pattern_create_mesh}()},
\code{\link{cairo_pattern_create_radial}()},
\code{\link{cairo_pattern_create_rgba}()},
\code{\link{cairo_pattern_create_rgb}()},
\code{\link{cairo_pattern_get_color_stop_count}()},
\code{\link{cairo_pattern_get_extend}()},
\code{\link{cairo_pattern_get_filter}()},
\code{\link{cairo_pattern_get_linear_points}()},
\code{\link{cairo_pattern_get_matrix}()},
\code{\link{cairo_pattern_get_radial_circles}()},
\code{\link{cairo_pattern_get_rgba}()},
\code{\link{cairo_pattern_get_type}()},
\code{\link{cairo_pattern_set_extend}()},
\code{\link{cairo_pattern_set_filter}()},
\code{\link{cairo_pattern_set_matrix}()},
\code{\link{cairo_pattern_status}()}
}
\concept{cairo-pattern}
