% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cairo_pattern_get_radial_circles}
\alias{cairo_pattern_get_radial_circles}
\title{cairo_pattern_get_radial_circles}
\usage{
cairo_pattern_get_radial_circles(pattern, x0, y0, r0, x1, y1, r1)
}
\arguments{
\item{pattern}{[\code{cairo_pattern_t *}] a #cairo_pattern_t}

\item{x0}{[\code{double *}] return value for the x coordinate of the center of the first circle, or \%NULL}

\item{y0}{[\code{double *}] return value for the y coordinate of the center of the first circle, or \%NULL}

\item{r0}{[\code{double *}] return value for the radius of the first circle, or \%NULL}

\item{x1}{[\code{double *}] return value for the x coordinate of the center of the second circle, or \%NULL}

\item{y1}{[\code{double *}] return value for the y coordinate of the center of the second circle, or \%NULL}

\item{r1}{[\code{double *}] return value for the radius of the second circle, or \%NULL}
}
\value{
[\code{int}]
}
\description{
Gets the gradient endpoint circles for a radial gradient, each
specified as a center coordinate and a radius.
}
\details{
Return value: \%CAIRO_STATUS_SUCCESS, or
\%CAIRO_STATUS_PATTERN_TYPE_MISMATCH if pattern is not a radial
gradient pattern.

Since: 1.4

C function prototype: \code{cairo_status_t cairo_pattern_get_radial_circles (cairo_pattern_t *pattern, double *x0, double *y0, double *r0, double *x1, double *y1, double *r1)}
}
\seealso{
Other cairo-pattern: 
\code{\link{cairo_mesh_pattern_begin_patch}()},
\code{\link{cairo_mesh_pattern_curve_to}()},
\code{\link{cairo_mesh_pattern_end_patch}()},
\code{\link{cairo_mesh_pattern_get_control_point}()},
\code{\link{cairo_mesh_pattern_get_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_get_patch_count}()},
\code{\link{cairo_mesh_pattern_get_path}()},
\code{\link{cairo_mesh_pattern_line_to}()},
\code{\link{cairo_mesh_pattern_move_to}()},
\code{\link{cairo_mesh_pattern_set_control_point}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgba}()},
\code{\link{cairo_mesh_pattern_set_corner_color_rgb}()},
\code{\link{cairo_pattern_add_color_stop_rgba}()},
\code{\link{cairo_pattern_add_color_stop_rgb}()},
\code{\link{cairo_pattern_create_for_surface}()},
\code{\link{cairo_pattern_create_linear}()},
\code{\link{cairo_pattern_create_mesh}()},
\code{\link{cairo_pattern_create_radial}()},
\code{\link{cairo_pattern_create_rgba}()},
\code{\link{cairo_pattern_create_rgb}()},
\code{\link{cairo_pattern_get_color_stop_count}()},
\code{\link{cairo_pattern_get_color_stop_rgba}()},
\code{\link{cairo_pattern_get_extend}()},
\code{\link{cairo_pattern_get_filter}()},
\code{\link{cairo_pattern_get_linear_points}()},
\code{\link{cairo_pattern_get_matrix}()},
\code{\link{cairo_pattern_get_rgba}()},
\code{\link{cairo_pattern_get_type}()},
\code{\link{cairo_pattern_set_extend}()},
\code{\link{cairo_pattern_set_filter}()},
\code{\link{cairo_pattern_set_matrix}()},
\code{\link{cairo_pattern_status}()}
}
\concept{cairo-pattern}
