% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_pattern_type_t}
\alias{cairo_pattern_type_t}
\title{cairo_pattern_type_t}
\format{
An object of class \code{list} of length 6.
}
\usage{
cairo_pattern_type_t
}
\description{
#cairo_pattern_type_t is used to describe the type of a given pattern.
}
\details{
The type of a pattern is determined by the function used to create
it. The cairo_pattern_create_rgb() and cairo_pattern_create_rgba()
functions create SOLID patterns. The remaining
cairo_pattern_create<!-- --> functions map to pattern types in obvious
ways.

The pattern type can be queried with cairo_pattern_get_type()

Most #cairo_pattern_t functions can be called with a pattern of any
type, (though trying to change the extend or filter for a solid
pattern will have no effect). A notable exception is
cairo_pattern_add_color_stop_rgb() and
cairo_pattern_add_color_stop_rgba() which must only be called with
gradient patterns (either LINEAR or RADIAL). Otherwise the pattern
will be shutdown and put into an error state.

New entries may be added in future versions.

Since: 1.2

\describe{
\item{CAIRO_PATTERN_TYPE_SOLID}{The pattern is a solid (uniform) color. It may be opaque or translucent, since 1.2.}
\item{CAIRO_PATTERN_TYPE_SURFACE}{The pattern is a based on a surface (an image), since 1.2.}
\item{CAIRO_PATTERN_TYPE_LINEAR}{The pattern is a linear gradient, since 1.2.}
\item{CAIRO_PATTERN_TYPE_RADIAL}{The pattern is a radial gradient, since 1.2.}
\item{CAIRO_PATTERN_TYPE_MESH}{The pattern is a mesh, since 1.12.}
\item{CAIRO_PATTERN_TYPE_RASTER_SOURCE}{The pattern is a user pattern providing raster data, since 1.12.}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
