% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-enums.R
\docType{data}
\name{cairo_pdf_metadata_t}
\alias{cairo_pdf_metadata_t}
\title{cairo_pdf_metadata_t}
\format{
An object of class \code{list} of length 7.
}
\usage{
cairo_pdf_metadata_t
}
\description{
#cairo_pdf_metadata_t is used by the
cairo_pdf_surface_set_metadata() function specify the metadata to set.
}
\details{
Since: 1.16

\describe{
\item{CAIRO_PDF_METADATA_TITLE}{The document title (Since 1.16)}
\item{CAIRO_PDF_METADATA_AUTHOR}{The document author (Since 1.16)}
\item{CAIRO_PDF_METADATA_SUBJECT}{The document subject (Since 1.16)}
\item{CAIRO_PDF_METADATA_KEYWORDS}{The document keywords (Since 1.16)}
\item{CAIRO_PDF_METADATA_CREATOR}{The document creator (Since 1.16)}
\item{CAIRO_PDF_METADATA_CREATE_DATE}{The document creation date (Since 1.16)}
\item{CAIRO_PDF_METADATA_MOD_DATE}{The document modification date (Since 1.16)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
