% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-enums.R
\docType{data}
\name{cairo_pdf_outline_flags_t}
\alias{cairo_pdf_outline_flags_t}
\title{cairo_pdf_outline_flags_t}
\format{
An object of class \code{list} of length 3.
}
\usage{
cairo_pdf_outline_flags_t
}
\description{
#cairo_pdf_outline_flags_t is used by the
cairo_pdf_surface_add_outline() function specify the attributes of
an outline item. These flags may be bitwise-or'd to produce any
combination of flags.
}
\details{
Since: 1.16

\describe{
\item{CAIRO_PDF_OUTLINE_FLAG_OPEN}{The outline item defaults to open in the PDF viewer (Since 1.16)}
\item{CAIRO_PDF_OUTLINE_FLAG_BOLD}{The outline item is displayed by the viewer in bold text (Since 1.16)}
\item{CAIRO_PDF_OUTLINE_FLAG_ITALIC}{The outline item is displayed by the viewer in italic text (Since 1.16)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
