% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_surface_restrict_to_version}
\alias{cairo_pdf_surface_restrict_to_version}
\title{cairo_pdf_surface_restrict_to_version}
\usage{
cairo_pdf_surface_restrict_to_version(surface, version)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PDF #cairo_surface_t}

\item{version}{[\code{int}] PDF version}
}
\description{
Restricts the generated PDF file to version. See cairo_pdf_get_versions()
for a list of available version values that can be used here.
}
\details{
This function should only be called before any drawing operations
have been performed on the given surface. The simplest way to do
this is to call this function immediately after creating the
surface.

Since: 1.10

C function prototype: \code{void cairo_pdf_surface_restrict_to_version (cairo_surface_t *surface, cairo_pdf_version_t version)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_set_metadata}()},
\code{\link{cairo_pdf_surface_set_page_label}()},
\code{\link{cairo_pdf_surface_set_size}()},
\code{\link{cairo_pdf_surface_set_thumbnail_size}()},
\code{\link{cairo_pdf_version_to_string}()}
}
\concept{cairo-pdf-surface}
