% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_surface_set_metadata}
\alias{cairo_pdf_surface_set_metadata}
\title{cairo_pdf_surface_set_metadata}
\usage{
cairo_pdf_surface_set_metadata(surface, metadata, utf8)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PDF #cairo_surface_t}

\item{metadata}{[\code{int}] The metadata item to set.}

\item{utf8}{[\code{char *}] metadata value}
}
\description{
Set document metadata. The \%CAIRO_PDF_METADATA_CREATE_DATE and
\%CAIRO_PDF_METADATA_MOD_DATE values must be in ISO-8601 format:
YYYY-MM-DDThh:mm:ss. An optional timezone of the form "[+/-]hh:mm"
or "Z" for UTC time can be appended. All other metadata values can be any UTF-8
string.
}
\details{
For example:
<informalexample><programlisting>
cairo_pdf_surface_set_metadata (surface, CAIRO_PDF_METADATA_TITLE, "My Document");
cairo_pdf_surface_set_metadata (surface, CAIRO_PDF_METADATA_CREATE_DATE, "2015-12-31T23:59+02:00");
</programlisting></informalexample>

Since: 1.16

C function prototype: \code{void cairo_pdf_surface_set_metadata (cairo_surface_t *surface, cairo_pdf_metadata_t metadata, const char *utf8)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_restrict_to_version}()},
\code{\link{cairo_pdf_surface_set_page_label}()},
\code{\link{cairo_pdf_surface_set_size}()},
\code{\link{cairo_pdf_surface_set_thumbnail_size}()},
\code{\link{cairo_pdf_version_to_string}()}
}
\concept{cairo-pdf-surface}
