% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_surface_set_page_label}
\alias{cairo_pdf_surface_set_page_label}
\title{cairo_pdf_surface_set_page_label}
\usage{
cairo_pdf_surface_set_page_label(surface, utf8)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PDF #cairo_surface_t}

\item{utf8}{[\code{char *}] The page label.}
}
\description{
Set page label for the current page.
}
\details{
Since: 1.16

C function prototype: \code{void cairo_pdf_surface_set_page_label (cairo_surface_t *surface, const char *utf8)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_restrict_to_version}()},
\code{\link{cairo_pdf_surface_set_metadata}()},
\code{\link{cairo_pdf_surface_set_size}()},
\code{\link{cairo_pdf_surface_set_thumbnail_size}()},
\code{\link{cairo_pdf_version_to_string}()}
}
\concept{cairo-pdf-surface}
