% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_surface_set_size}
\alias{cairo_pdf_surface_set_size}
\title{cairo_pdf_surface_set_size}
\usage{
cairo_pdf_surface_set_size(surface, width_in_points, height_in_points)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PDF #cairo_surface_t}

\item{width_in_points}{[\code{double}] new surface width, in points (1 point == 1/72.0 inch)}

\item{height_in_points}{[\code{double}] new surface height, in points (1 point == 1/72.0 inch)}
}
\description{
Changes the size of a PDF surface for the current (and
subsequent) pages.
}
\details{
This function should only be called before any drawing operations
have been performed on the current page. The simplest way to do
this is to call this function immediately after creating the
surface or immediately after completing a page with either
cairo_show_page() or cairo_copy_page().

Since: 1.2

C function prototype: \code{void cairo_pdf_surface_set_size (cairo_surface_t *surface, double width_in_points, double height_in_points)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_restrict_to_version}()},
\code{\link{cairo_pdf_surface_set_metadata}()},
\code{\link{cairo_pdf_surface_set_page_label}()},
\code{\link{cairo_pdf_surface_set_thumbnail_size}()},
\code{\link{cairo_pdf_version_to_string}()}
}
\concept{cairo-pdf-surface}
