% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_surface_set_thumbnail_size}
\alias{cairo_pdf_surface_set_thumbnail_size}
\title{cairo_pdf_surface_set_thumbnail_size}
\usage{
cairo_pdf_surface_set_thumbnail_size(surface, width, height)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PDF #cairo_surface_t}

\item{width}{[\code{int}] Thumbnail width.}

\item{height}{[\code{int}] Thumbnail height}
}
\description{
Set the thumbnail image size for the current and all subsequent
pages. Setting a width or height of 0 disables thumbnails for the
current and subsequent pages.
}
\details{
Since: 1.16

C function prototype: \code{void cairo_pdf_surface_set_thumbnail_size (cairo_surface_t *surface, int width, int height)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_restrict_to_version}()},
\code{\link{cairo_pdf_surface_set_metadata}()},
\code{\link{cairo_pdf_surface_set_page_label}()},
\code{\link{cairo_pdf_surface_set_size}()},
\code{\link{cairo_pdf_version_to_string}()}
}
\concept{cairo-pdf-surface}
