% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pdf-surface.R
\name{cairo_pdf_version_to_string}
\alias{cairo_pdf_version_to_string}
\title{cairo_pdf_version_to_string}
\usage{
cairo_pdf_version_to_string(version)
}
\arguments{
\item{version}{[\code{int}] a version id}
}
\value{
[\code{const char *}]
}
\description{
Get the string representation of the given version id. This function
will return \%NULL if version isn't valid. See cairo_pdf_get_versions()
for a way to get the list of valid version ids.
}
\details{
Return value: the string associated to given version.

Since: 1.10

C function prototype: \code{const char * cairo_pdf_version_to_string (cairo_pdf_version_t version)}
}
\seealso{
Other cairo-pdf-surface: 
\code{\link{cairo_pdf_surface_add_outline}()},
\code{\link{cairo_pdf_surface_create}()},
\code{\link{cairo_pdf_surface_restrict_to_version}()},
\code{\link{cairo_pdf_surface_set_metadata}()},
\code{\link{cairo_pdf_surface_set_page_label}()},
\code{\link{cairo_pdf_surface_set_size}()},
\code{\link{cairo_pdf_surface_set_thumbnail_size}()}
}
\concept{cairo-pdf-surface}
