% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--bespoke-polygon.R
\name{cairo_polygon}
\alias{cairo_polygon}
\alias{cairo_polyline}
\title{Polygon, Polyline}
\usage{
cairo_polygon(cr, x, y)

cairo_polyline(cr, x, y)
}
\arguments{
\item{cr}{\code{cairo_t *}}

\item{x}{\code{double}}

\item{y}{\code{double}}
}
\description{
Bespoke
}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}

Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}
}
\concept{bespoke}
