% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--bespoke.R
\name{cairo_polygon_vec}
\alias{cairo_polygon_vec}
\title{Draw lots of polygons}
\usage{
cairo_polygon_vec(cr, x, y, idx, r, g, b, a, sr, sg, sb, sa)
}
\arguments{
\item{cr}{\code{cairo_t *}}

\item{x}{\code{double}}

\item{y}{\code{double}}

\item{idx}{\code{integer} runs of the same value define the locations in the
x and y vectors which refer to the same polygon.}

\item{r, g, b, a}{fill colour values in range [0,1]. Set \code{a = 0} to
disable filling.}

\item{sr, sg, sb, sa}{stroke colour values in range [0,1]. Set \code{sa = 0} to
disable drawing the stroke}
}
\description{
This is a custom function to draw lots of polygons within C with a single
call from R.  Use this function when you want to draw lots of polygon and
need to speed up the creation of the final image.
}
\details{
The x and y arguments must be the same length.

The colour arguments (r, g, b, a, etc) must all be the same length as the
number of distinct runs of values in \code{idx}, or length of 1
}
\examples{
\dontrun{
# define 2 polygons
df <- tibble::tribble(
~x, ~y, ~idx,
 0,  0,    1,
 1,  0,    1,
 1,  1,    1,
 0,  1,    1,
10,  0,   55,
11,  0,   55,
11,  1,   55,
10,  1,   55
)

# Draw 2 polygons with red fill and blue border
cairo_polygon_vec(cr, df$x, df$y, df$idx,  1, 0, 0, 1,   0, 0, 1, 1)

}


}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}

Other vectorised: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_rectangle_vec}()},
\code{\link{cairo_segment_vec}()}
}
\concept{bespoke}
\concept{vectorised}
