% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps-enums.R
\docType{data}
\name{cairo_ps_level_t}
\alias{cairo_ps_level_t}
\title{cairo_ps_level_t}
\format{
An object of class \code{list} of length 2.
}
\usage{
cairo_ps_level_t
}
\description{
#cairo_ps_level_t is used to describe the language level of the
PostScript Language Reference that a generated PostScript file will
conform to.
}
\details{
Since: 1.6

\describe{
\item{CAIRO_PS_LEVEL_2}{The language level 2 of the PostScript specification. (Since 1.6)}
\item{CAIRO_PS_LEVEL_3}{The language level 3 of the PostScript specification. (Since 1.6)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
