% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_level_to_string}
\alias{cairo_ps_level_to_string}
\title{cairo_ps_level_to_string}
\usage{
cairo_ps_level_to_string(level)
}
\arguments{
\item{level}{[\code{int}] a level id}
}
\value{
[\code{const char *}]
}
\description{
Get the string representation of the given level id. This function
will return \%NULL if level id isn't valid. See cairo_ps_get_levels()
for a way to get the list of valid level ids.
}
\details{
Return value: the string associated to given level.

Since: 1.6

C function prototype: \code{const char * cairo_ps_level_to_string (cairo_ps_level_t level)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_page_setup}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_dsc_comment}()},
\code{\link{cairo_ps_surface_get_eps}()},
\code{\link{cairo_ps_surface_restrict_to_level}()},
\code{\link{cairo_ps_surface_set_eps}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
