% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_surface_dsc_begin_page_setup}
\alias{cairo_ps_surface_dsc_begin_page_setup}
\title{cairo_ps_surface_dsc_begin_page_setup}
\usage{
cairo_ps_surface_dsc_begin_page_setup(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PostScript #cairo_surface_t}
}
\description{
This function indicates that subsequent calls to
cairo_ps_surface_dsc_comment() should direct comments to the
PageSetup section of the PostScript output.
}
\details{
This function call is only needed for the first page of a
surface. It should be called after any call to
cairo_ps_surface_dsc_begin_setup() and before any drawing is
performed to the surface.

See cairo_ps_surface_dsc_comment() for more details.

Since: 1.2

C function prototype: \code{void cairo_ps_surface_dsc_begin_page_setup (cairo_surface_t *surface)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_level_to_string}()},
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_dsc_comment}()},
\code{\link{cairo_ps_surface_get_eps}()},
\code{\link{cairo_ps_surface_restrict_to_level}()},
\code{\link{cairo_ps_surface_set_eps}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
