% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_surface_dsc_comment}
\alias{cairo_ps_surface_dsc_comment}
\title{cairo_ps_surface_dsc_comment}
\usage{
cairo_ps_surface_dsc_comment(surface, comment)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PostScript #cairo_surface_t}

\item{comment}{[\code{char *}] a comment string to be emitted into the PostScript output}
}
\description{
Emit a comment into the PostScript output for the given surface.
}
\details{
The comment is expected to conform to the PostScript Language
Document Structuring Conventions (DSC). Please see that manual for
details on the available comments and their meanings. In
particular, the \\%\\%IncludeFeature comment allows a
device-independent means of controlling printer device features. So
the PostScript Printer Description Files Specification will also be
a useful reference.

The comment string must begin with a percent character (\\%) and the
total length of the string (including any initial percent
characters) must not exceed 255 characters. Violating either of
these conditions will place surface into an error state. But
beyond these two conditions, this function will not enforce
conformance of the comment with any particular specification.

The comment string should not have a trailing newline.

The DSC specifies different sections in which particular comments
can appear. This function provides for comments to be emitted
within three sections: the header, the Setup section, and the
PageSetup section.  Comments appearing in the first two sections
apply to the entire document while comments in the BeginPageSetup
section apply only to a single page.

For comments to appear in the header section, this function should
be called after the surface is created, but before a call to
cairo_ps_surface_dsc_begin_setup().

For comments to appear in the Setup section, this function should
be called after a call to cairo_ps_surface_dsc_begin_setup() but
before a call to cairo_ps_surface_dsc_begin_page_setup().

For comments to appear in the PageSetup section, this function
should be called after a call to
cairo_ps_surface_dsc_begin_page_setup().

Note that it is only necessary to call
cairo_ps_surface_dsc_begin_page_setup() for the first page of any
surface. After a call to cairo_show_page() or cairo_copy_page()
comments are unambiguously directed to the PageSetup section of the
current page. But it doesn't hurt to call this function at the
beginning of every page as that consistency may make the calling
code simpler.

As a final note, cairo automatically generates several comments on
its own. As such, applications must not manually generate any of
the following comments:

Header section: \\%!PS-Adobe-3.0, \\%\\%Creator, \\%\\%CreationDate, \\%\\%Pages,
\\%\\%BoundingBox, \\%\\%DocumentData, \\%\\%LanguageLevel, \\%\\%EndComments.

Setup section: \\%\\%BeginSetup, \\%\\%EndSetup

PageSetup section: \\%\\%BeginPageSetup, \\%\\%PageBoundingBox, \\%\\%EndPageSetup.

Other sections: \\%\\%BeginProlog, \\%\\%EndProlog, \\%\\%Page, \\%\\%Trailer, \\%\\%EOF

Here is an example sequence showing how this function might be used:

<informalexample><programlisting>
cairo_surface_t *surface = cairo_ps_surface_create (filename, width, height);
...
cairo_ps_surface_dsc_comment (surface, "\%\%Title: My excellent document");
cairo_ps_surface_dsc_comment (surface, "\%\%Copyright: Copyright (C) 2006 Cairo Lover")
...
cairo_ps_surface_dsc_begin_setup (surface);
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *MediaColor White");
...
cairo_ps_surface_dsc_begin_page_setup (surface);
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *PageSize A3");
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *InputSlot LargeCapacity");
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *MediaType Glossy");
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *MediaColor Blue");
... draw to first page here ..
cairo_show_page (cr);
...
cairo_ps_surface_dsc_comment (surface, "\%\%IncludeFeature: *PageSize A5");
...
</programlisting></informalexample>

Since: 1.2

C function prototype: \code{void cairo_ps_surface_dsc_comment (cairo_surface_t *surface, const char *comment)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_level_to_string}()},
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_page_setup}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_get_eps}()},
\code{\link{cairo_ps_surface_restrict_to_level}()},
\code{\link{cairo_ps_surface_set_eps}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
