% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_surface_get_eps}
\alias{cairo_ps_surface_get_eps}
\title{cairo_ps_surface_get_eps}
\usage{
cairo_ps_surface_get_eps(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PostScript #cairo_surface_t}
}
\value{
[\code{int}]
}
\description{
Check whether the PostScript surface will output Encapsulated PostScript.
}
\details{
Return value: \%TRUE if the surface will output Encapsulated PostScript.

Since: 1.6

C function prototype: \code{cairo_public cairo_bool_t cairo_ps_surface_get_eps (cairo_surface_t *surface)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_level_to_string}()},
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_page_setup}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_dsc_comment}()},
\code{\link{cairo_ps_surface_restrict_to_level}()},
\code{\link{cairo_ps_surface_set_eps}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
