% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_surface_restrict_to_level}
\alias{cairo_ps_surface_restrict_to_level}
\title{cairo_ps_surface_restrict_to_level}
\usage{
cairo_ps_surface_restrict_to_level(surface, level)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PostScript #cairo_surface_t}

\item{level}{[\code{int}] PostScript level}
}
\description{
Restricts the generated PostSript file to level. See
cairo_ps_get_levels() for a list of available level values that
can be used here.
}
\details{
This function should only be called before any drawing operations
have been performed on the given surface. The simplest way to do
this is to call this function immediately after creating the
surface.

Since: 1.6

C function prototype: \code{void cairo_ps_surface_restrict_to_level (cairo_surface_t *surface, cairo_ps_level_t level)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_level_to_string}()},
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_page_setup}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_dsc_comment}()},
\code{\link{cairo_ps_surface_get_eps}()},
\code{\link{cairo_ps_surface_set_eps}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
