% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-ps.R
\name{cairo_ps_surface_set_eps}
\alias{cairo_ps_surface_set_eps}
\title{cairo_ps_surface_set_eps}
\usage{
cairo_ps_surface_set_eps(surface, eps)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a PostScript #cairo_surface_t}

\item{eps}{[\code{int}] \%TRUE to output EPS format PostScript}
}
\description{
If eps is \%TRUE, the PostScript surface will output Encapsulated
PostScript.
}
\details{
This function should only be called before any drawing operations
have been performed on the current page. The simplest way to do
this is to call this function immediately after creating the
surface. An Encapsulated PostScript file should never contain more
than one page.

Since: 1.6

C function prototype: \code{void cairo_ps_surface_set_eps (cairo_surface_t *surface, cairo_bool_t eps)}
}
\seealso{
Other cairo-ps: 
\code{\link{cairo_ps_level_to_string}()},
\code{\link{cairo_ps_surface_create}()},
\code{\link{cairo_ps_surface_dsc_begin_page_setup}()},
\code{\link{cairo_ps_surface_dsc_begin_setup}()},
\code{\link{cairo_ps_surface_dsc_comment}()},
\code{\link{cairo_ps_surface_get_eps}()},
\code{\link{cairo_ps_surface_restrict_to_level}()},
\code{\link{cairo_ps_surface_set_size}()}
}
\concept{cairo-ps}
