% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--.R
\name{cairo_rectangle}
\alias{cairo_rectangle}
\title{cairo_rectangle}
\usage{
cairo_rectangle(cr, x, y, width, height)
}
\arguments{
\item{cr}{[\code{cairo_t *}] a cairo context}

\item{x}{[\code{double}] the X coordinate of the top left corner of the rectangle}

\item{y}{[\code{double}] the Y coordinate to the top left corner of the rectangle}

\item{width}{[\code{double}] the width of the rectangle}

\item{height}{[\code{double}] the height of the rectangle}
}
\description{
Adds a closed sub-path rectangle of the given size to the current
path at position (x, y) in user-space coordinates.
}
\details{
This function is logically equivalent to:
<informalexample><programlisting>
cairo_move_to (cr, x, y);
cairo_rel_line_to (cr, width, 0);
cairo_rel_line_to (cr, 0, height);
cairo_rel_line_to (cr, -width, 0);
cairo_close_path (cr);
</programlisting></informalexample>

Since: 1.0

C function prototype: \code{void cairo_rectangle (cairo_t *cr, double x, double y, double width, double height)}
}
\seealso{
Other cairo--: 
\code{\link{cairo_append_path}()},
\code{\link{cairo_arc_negative}()},
\code{\link{cairo_arc}()},
\code{\link{cairo_clip_extents}()},
\code{\link{cairo_clip_preserve}()},
\code{\link{cairo_clip}()},
\code{\link{cairo_close_path}()},
\code{\link{cairo_copy_page}()},
\code{\link{cairo_copy_path_flat}()},
\code{\link{cairo_copy_path}()},
\code{\link{cairo_create}()},
\code{\link{cairo_curve_to}()},
\code{\link{cairo_device_to_user_distance}()},
\code{\link{cairo_device_to_user}()},
\code{\link{cairo_fill_extents}()},
\code{\link{cairo_fill_preserve}()},
\code{\link{cairo_fill}()},
\code{\link{cairo_font_extents}()},
\code{\link{cairo_get_antialias}()},
\code{\link{cairo_get_current_point}()},
\code{\link{cairo_get_dash_count}()},
\code{\link{cairo_get_dash}()},
\code{\link{cairo_get_fill_rule}()},
\code{\link{cairo_get_font_face}()},
\code{\link{cairo_get_font_matrix}()},
\code{\link{cairo_get_group_target}()},
\code{\link{cairo_get_line_cap}()},
\code{\link{cairo_get_line_join}()},
\code{\link{cairo_get_line_width}()},
\code{\link{cairo_get_matrix}()},
\code{\link{cairo_get_miter_limit}()},
\code{\link{cairo_get_operator}()},
\code{\link{cairo_get_source}()},
\code{\link{cairo_get_target}()},
\code{\link{cairo_get_tolerance}()},
\code{\link{cairo_has_current_point}()},
\code{\link{cairo_identity_matrix}()},
\code{\link{cairo_in_clip}()},
\code{\link{cairo_in_fill}()},
\code{\link{cairo_in_stroke}()},
\code{\link{cairo_line_to}()},
\code{\link{cairo_mask_surface}()},
\code{\link{cairo_mask}()},
\code{\link{cairo_move_to}()},
\code{\link{cairo_new_path}()},
\code{\link{cairo_new_sub_path}()},
\code{\link{cairo_paint_with_alpha}()},
\code{\link{cairo_paint}()},
\code{\link{cairo_path_extents}()},
\code{\link{cairo_pop_group_to_source}()},
\code{\link{cairo_pop_group}()},
\code{\link{cairo_push_group_with_content}()},
\code{\link{cairo_push_group}()},
\code{\link{cairo_rel_curve_to}()},
\code{\link{cairo_rel_line_to}()},
\code{\link{cairo_rel_move_to}()},
\code{\link{cairo_reset_clip}()},
\code{\link{cairo_restore}()},
\code{\link{cairo_rotate}()},
\code{\link{cairo_save}()},
\code{\link{cairo_scale}()},
\code{\link{cairo_select_font_face}()},
\code{\link{cairo_set_antialias}()},
\code{\link{cairo_set_dash}()},
\code{\link{cairo_set_fill_rule}()},
\code{\link{cairo_set_font_face}()},
\code{\link{cairo_set_font_matrix}()},
\code{\link{cairo_set_font_size}()},
\code{\link{cairo_set_line_cap}()},
\code{\link{cairo_set_line_join}()},
\code{\link{cairo_set_line_width}()},
\code{\link{cairo_set_matrix}()},
\code{\link{cairo_set_miter_limit}()},
\code{\link{cairo_set_operator}()},
\code{\link{cairo_set_source_rgba}()},
\code{\link{cairo_set_source_rgb}()},
\code{\link{cairo_set_source_surface}()},
\code{\link{cairo_set_source}()},
\code{\link{cairo_set_tolerance}()},
\code{\link{cairo_show_page}()},
\code{\link{cairo_show_text}()},
\code{\link{cairo_status}()},
\code{\link{cairo_stroke_extents}()},
\code{\link{cairo_stroke_preserve}()},
\code{\link{cairo_stroke}()},
\code{\link{cairo_tag_begin}()},
\code{\link{cairo_tag_end}()},
\code{\link{cairo_text_extents}()},
\code{\link{cairo_text_path}()},
\code{\link{cairo_transform}()},
\code{\link{cairo_translate}()},
\code{\link{cairo_user_to_device_distance}()},
\code{\link{cairo_user_to_device}()}
}
\concept{cairo--}
