% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--structs.R
\name{cairo_rectangle_t}
\alias{cairo_rectangle_t}
\title{cairo_rectangle_t}
\usage{
cairo_rectangle_t(x, y, width, height)
}
\arguments{
\item{x}{X coordinate of the left side of the rectangle}

\item{y}{Y coordinate of the the top side of the rectangle}

\item{width}{width of the rectangle}

\item{height}{height of the rectangle}
}
\description{
A data structure for holding a rectangle.
}
\details{
Since: 1.4
}
\seealso{
Other struct: 
\code{\link{as.list.cairo_font_extents_t}()},
\code{\link{as.list.cairo_matrix_t}()},
\code{\link{as.list.cairo_rectangle_t}()},
\code{\link{as.list.cairo_text_extents_t}()},
\code{\link{cairo_font_extents_t}()},
\code{\link{cairo_matrix_t_update}()},
\code{\link{cairo_matrix_t}()},
\code{\link{cairo_rectangle_int_t}()},
\code{\link{cairo_text_extents_t}()}
}
\concept{struct}
