% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--.R
\name{cairo_rel_curve_to}
\alias{cairo_rel_curve_to}
\title{cairo_rel_curve_to}
\usage{
cairo_rel_curve_to(cr, dx1, dy1, dx2, dy2, dx3, dy3)
}
\arguments{
\item{cr}{[\code{cairo_t *}] a cairo context}

\item{dx1}{[\code{double}] the X offset to the first control point}

\item{dy1}{[\code{double}] the Y offset to the first control point}

\item{dx2}{[\code{double}] the X offset to the second control point}

\item{dy2}{[\code{double}] the Y offset to the second control point}

\item{dx3}{[\code{double}] the X offset to the end of the curve}

\item{dy3}{[\code{double}] the Y offset to the end of the curve}
}
\description{
Relative-coordinate version of cairo_curve_to(). All offsets are
relative to the current point. Adds a cubic Bézier spline to the
path from the current point to a point offset from the current
point by (dx3, dy3), using points offset by (dx1, dy1) and
(dx2, dy2) as the control points. After this call the current
point will be offset by (dx3, dy3).
}
\details{
Given a current point of (x, y), cairo_rel_curve_to(cr, dx1,
dy1, dx2, dy2, dx3, dy3) is logically equivalent to
cairo_curve_to(cr, x+dx1, y+dy1, x+dx2, y+dy2, x+dx3, y+dy3).

It is an error to call this function with no current point. Doing
so will cause cr to shutdown with a status of
\%CAIRO_STATUS_NO_CURRENT_POINT.

Since: 1.0

C function prototype: \code{void cairo_rel_curve_to (cairo_t *cr, double dx1, double dy1, double dx2, double dy2, double dx3, double dy3)}
}
\seealso{
Other cairo--: 
\code{\link{cairo_append_path}()},
\code{\link{cairo_arc_negative}()},
\code{\link{cairo_arc}()},
\code{\link{cairo_clip_extents}()},
\code{\link{cairo_clip_preserve}()},
\code{\link{cairo_clip}()},
\code{\link{cairo_close_path}()},
\code{\link{cairo_copy_page}()},
\code{\link{cairo_copy_path_flat}()},
\code{\link{cairo_copy_path}()},
\code{\link{cairo_create}()},
\code{\link{cairo_curve_to}()},
\code{\link{cairo_device_to_user_distance}()},
\code{\link{cairo_device_to_user}()},
\code{\link{cairo_fill_extents}()},
\code{\link{cairo_fill_preserve}()},
\code{\link{cairo_fill}()},
\code{\link{cairo_font_extents}()},
\code{\link{cairo_get_antialias}()},
\code{\link{cairo_get_current_point}()},
\code{\link{cairo_get_dash_count}()},
\code{\link{cairo_get_dash}()},
\code{\link{cairo_get_fill_rule}()},
\code{\link{cairo_get_font_face}()},
\code{\link{cairo_get_font_matrix}()},
\code{\link{cairo_get_group_target}()},
\code{\link{cairo_get_line_cap}()},
\code{\link{cairo_get_line_join}()},
\code{\link{cairo_get_line_width}()},
\code{\link{cairo_get_matrix}()},
\code{\link{cairo_get_miter_limit}()},
\code{\link{cairo_get_operator}()},
\code{\link{cairo_get_source}()},
\code{\link{cairo_get_target}()},
\code{\link{cairo_get_tolerance}()},
\code{\link{cairo_has_current_point}()},
\code{\link{cairo_identity_matrix}()},
\code{\link{cairo_in_clip}()},
\code{\link{cairo_in_fill}()},
\code{\link{cairo_in_stroke}()},
\code{\link{cairo_line_to}()},
\code{\link{cairo_mask_surface}()},
\code{\link{cairo_mask}()},
\code{\link{cairo_move_to}()},
\code{\link{cairo_new_path}()},
\code{\link{cairo_new_sub_path}()},
\code{\link{cairo_paint_with_alpha}()},
\code{\link{cairo_paint}()},
\code{\link{cairo_path_extents}()},
\code{\link{cairo_pop_group_to_source}()},
\code{\link{cairo_pop_group}()},
\code{\link{cairo_push_group_with_content}()},
\code{\link{cairo_push_group}()},
\code{\link{cairo_rectangle}()},
\code{\link{cairo_rel_line_to}()},
\code{\link{cairo_rel_move_to}()},
\code{\link{cairo_reset_clip}()},
\code{\link{cairo_restore}()},
\code{\link{cairo_rotate}()},
\code{\link{cairo_save}()},
\code{\link{cairo_scale}()},
\code{\link{cairo_select_font_face}()},
\code{\link{cairo_set_antialias}()},
\code{\link{cairo_set_dash}()},
\code{\link{cairo_set_fill_rule}()},
\code{\link{cairo_set_font_face}()},
\code{\link{cairo_set_font_matrix}()},
\code{\link{cairo_set_font_size}()},
\code{\link{cairo_set_line_cap}()},
\code{\link{cairo_set_line_join}()},
\code{\link{cairo_set_line_width}()},
\code{\link{cairo_set_matrix}()},
\code{\link{cairo_set_miter_limit}()},
\code{\link{cairo_set_operator}()},
\code{\link{cairo_set_source_rgba}()},
\code{\link{cairo_set_source_rgb}()},
\code{\link{cairo_set_source_surface}()},
\code{\link{cairo_set_source}()},
\code{\link{cairo_set_tolerance}()},
\code{\link{cairo_show_page}()},
\code{\link{cairo_show_text}()},
\code{\link{cairo_status}()},
\code{\link{cairo_stroke_extents}()},
\code{\link{cairo_stroke_preserve}()},
\code{\link{cairo_stroke}()},
\code{\link{cairo_tag_begin}()},
\code{\link{cairo_tag_end}()},
\code{\link{cairo_text_extents}()},
\code{\link{cairo_text_path}()},
\code{\link{cairo_transform}()},
\code{\link{cairo_translate}()},
\code{\link{cairo_user_to_device_distance}()},
\code{\link{cairo_user_to_device}()}
}
\concept{cairo--}
