% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-script-enums.R
\docType{data}
\name{cairo_script_mode_t}
\alias{cairo_script_mode_t}
\title{cairo_script_mode_t}
\format{
An object of class \code{list} of length 2.
}
\usage{
cairo_script_mode_t
}
\description{
A set of script output variants.
}
\details{
Since: 1.12

\describe{
\item{CAIRO_SCRIPT_MODE_ASCII}{the output will be in readable text (default). (Since 1.12)}
\item{CAIRO_SCRIPT_MODE_BINARY}{the output will use byte codes. (Since 1.12)}
}
}
\keyword{datasets}
