% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--bespoke.R
\name{cairo_segment_vec}
\alias{cairo_segment_vec}
\title{Draw lots of line segments}
\usage{
cairo_segment_vec(cr, x1, y1, x2, y2, r, g, b, a)
}
\arguments{
\item{cr}{\code{cairo_t *}}

\item{x1, y1, x2, y2}{coordinates of endpoints}

\item{r, g, b, a}{colour values in range [0,1]}
}
\description{
This is a custom function to draw lots of line segments within C with a single
call from R.  Use this function when you want to draw lots of segments and
need to speed up the creation of the final image.
}
\details{
Besides the \code{cr} context, all arguments should either be the same length
as the initial argument (\code{xc}) or have a length of 1.

Vectors which have a length of 1 will be considered to have a constant value
for all objects
}
\examples{
\dontrun{
# Draw 10 lines - all black
N <- 10
x1 <- runif(N)
y1 <- runif(N)
x2 <- runif(N)
y2 <- runif(N)
cairo_segment_vec(cr, x1, y2, x2, y2,  0, 0, 0, 1)
}


}
\seealso{
Other bespoke: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_image_surface_create_from_array}()},
\code{\link{cairo_image_surface_get_array}()},
\code{\link{cairo_image_surface_get_raster}()},
\code{\link{cairo_image_surface_get_raw}()},
\code{\link{cairo_image_surface_write_array}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_polygon}()},
\code{\link{cairo_rectangle_vec}()}

Other vectorised: 
\code{\link{cairo_arc_vec}()},
\code{\link{cairo_polygon_vec}()},
\code{\link{cairo_rectangle_vec}()}
}
\concept{bespoke}
\concept{vectorised}
