% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-surface.R
\name{cairo_surface_finish}
\alias{cairo_surface_finish}
\title{cairo_surface_finish}
\usage{
cairo_surface_finish(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] the #cairo_surface_t to finish}
}
\description{
This function finishes the surface and drops all references to
external resources.  For example, for the Xlib backend it means
that cairo will no longer access the drawable, which can be freed.
After calling cairo_surface_finish() the only valid operations on a
surface are checking status, getting and setting user, referencing
and destroying, and flushing and finishing it.
Further drawing to the surface will not affect the
surface but will instead trigger a \%CAIRO_STATUS_SURFACE_FINISHED
error.
}
\details{
When the last call to cairo_surface_destroy() decreases the
reference count to zero, cairo will call cairo_surface_finish() if
it hasn't been called already, before freeing the resources
associated with the surface.

Since: 1.0

C function prototype: \code{void cairo_surface_finish (cairo_surface_t *surface)}
}
\seealso{
Other cairo-surface: 
\code{\link{cairo_surface_flush}()}
}
\concept{cairo-surface}
