% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-surface.R
\name{cairo_surface_flush}
\alias{cairo_surface_flush}
\title{cairo_surface_flush}
\usage{
cairo_surface_flush(surface)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a #cairo_surface_t}
}
\description{
Do any pending drawing for the surface and also restore any temporary
modifications cairo has made to the surface's state. This function
must be called before switching from drawing on the surface with
cairo to drawing on it directly with native APIs, or accessing its
memory outside of Cairo. If the surface doesn't support direct
access, then this function does nothing.
}
\details{
Since: 1.0

C function prototype: \code{void cairo_surface_flush (cairo_surface_t *surface)}
}
\seealso{
Other cairo-surface: 
\code{\link{cairo_surface_finish}()}
}
\concept{cairo-surface}
