% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo--enum.R
\docType{data}
\name{cairo_surface_type_t}
\alias{cairo_surface_type_t}
\title{cairo_surface_type_t}
\format{
An object of class \code{list} of length 25.
}
\usage{
cairo_surface_type_t
}
\description{
#cairo_surface_type_t is used to describe the type of a given
surface. The surface types are also known as "backends" or "surface
backends" within cairo.
}
\details{
The type of a surface is determined by the function used to create
it, which will generally be of the form
<function>cairo_<emphasis>type</emphasis>_surface_create(<!-- -->)</function>,
(though see cairo_surface_create_similar() as well).

The surface type can be queried with cairo_surface_get_type()

The various #cairo_surface_t functions can be used with surfaces of
any type, but some backends also provide type-specific functions
that must only be called with a surface of the appropriate
type. These functions have names that begin with
<literal>cairo_<emphasis>type</emphasis>_surface</literal> such as cairo_image_surface_get_width().

The behavior of calling a type-specific function with a surface of
the wrong type is undefined.

New entries may be added in future versions.

Since: 1.2

\describe{
\item{CAIRO_SURFACE_TYPE_IMAGE}{The surface is of type image, since 1.2}
\item{CAIRO_SURFACE_TYPE_PDF}{The surface is of type pdf, since 1.2}
\item{CAIRO_SURFACE_TYPE_PS}{The surface is of type ps, since 1.2}
\item{CAIRO_SURFACE_TYPE_XLIB}{The surface is of type xlib, since 1.2}
\item{CAIRO_SURFACE_TYPE_XCB}{The surface is of type xcb, since 1.2}
\item{CAIRO_SURFACE_TYPE_GLITZ}{The surface is of type glitz, since 1.2}
\item{CAIRO_SURFACE_TYPE_QUARTZ}{The surface is of type quartz, since 1.2}
\item{CAIRO_SURFACE_TYPE_WIN32}{The surface is of type win32, since 1.2}
\item{CAIRO_SURFACE_TYPE_BEOS}{The surface is of type beos, since 1.2}
\item{CAIRO_SURFACE_TYPE_DIRECTFB}{The surface is of type directfb, since 1.2}
\item{CAIRO_SURFACE_TYPE_SVG}{The surface is of type svg, since 1.2}
\item{CAIRO_SURFACE_TYPE_OS2}{The surface is of type os2, since 1.4}
\item{CAIRO_SURFACE_TYPE_WIN32_PRINTING}{The surface is a win32 printing surface, since 1.6}
\item{CAIRO_SURFACE_TYPE_QUARTZ_IMAGE}{The surface is of type quartz_image, since 1.6}
\item{CAIRO_SURFACE_TYPE_SCRIPT}{The surface is of type script, since 1.10}
\item{CAIRO_SURFACE_TYPE_QT}{The surface is of type Qt, since 1.10}
\item{CAIRO_SURFACE_TYPE_RECORDING}{The surface is of type recording, since 1.10}
\item{CAIRO_SURFACE_TYPE_VG}{The surface is a OpenVG surface, since 1.10}
\item{CAIRO_SURFACE_TYPE_GL}{The surface is of type OpenGL, since 1.10}
\item{CAIRO_SURFACE_TYPE_DRM}{The surface is of type Direct Render Manager, since 1.10}
\item{CAIRO_SURFACE_TYPE_TEE}{The surface is of type 'tee' (a multiplexing surface), since 1.10}
\item{CAIRO_SURFACE_TYPE_XML}{The surface is of type XML (for debugging), since 1.10}
\item{CAIRO_SURFACE_TYPE_SUBSURFACE}{The surface is a subsurface created with cairo_surface_create_for_rectangle(), since 1.10}
\item{CAIRO_SURFACE_TYPE_COGL}{This surface is of type Cogl, since 1.12}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
