% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-png.R
\name{cairo_surface_write_to_png}
\alias{cairo_surface_write_to_png}
\title{cairo_surface_write_to_png}
\usage{
cairo_surface_write_to_png(surface, filename)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a #cairo_surface_t with pixel contents}

\item{filename}{[\code{char *}] the name of a file to write to; on Windows this filename is encoded in UTF-8.}
}
\value{
[\code{int}]
}
\description{
Writes the contents of surface to a new file filename as a PNG
image.
}
\details{
Return value: \%CAIRO_STATUS_SUCCESS if the PNG file was written
successfully. Otherwise, \%CAIRO_STATUS_NO_MEMORY if memory could not
be allocated for the operation or
\%CAIRO_STATUS_SURFACE_TYPE_MISMATCH if the surface does not have
pixel contents, or \%CAIRO_STATUS_WRITE_ERROR if an I/O error occurs
while attempting to write the file, or \%CAIRO_STATUS_PNG_ERROR if libpng
returned an error.

Since: 1.0

C function prototype: \code{cairo_status_t cairo_surface_write_to_png (cairo_surface_t *surface, const char *filename)}
}
\seealso{
Other cairo-png: 
\code{\link{cairo_image_surface_create_from_png}()}
}
\concept{cairo-png}
