% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-svg-surface.R
\name{cairo_svg_surface_restrict_to_version}
\alias{cairo_svg_surface_restrict_to_version}
\title{cairo_svg_surface_restrict_to_version}
\usage{
cairo_svg_surface_restrict_to_version(surface, version)
}
\arguments{
\item{surface}{[\code{cairo_surface_t *}] a SVG #cairo_surface_t}

\item{version}{[\code{int}] SVG version}
}
\description{
Restricts the generated SVG file to version. See cairo_svg_get_versions()
for a list of available version values that can be used here.
}
\details{
This function should only be called before any drawing operations
have been performed on the given surface. The simplest way to do
this is to call this function immediately after creating the
surface.

Since: 1.2

C function prototype: \code{void cairo_svg_surface_restrict_to_version (cairo_surface_t *surface, cairo_svg_version_t version)}
}
\seealso{
Other cairo-svg-surface: 
\code{\link{cairo_svg_surface_create}()},
\code{\link{cairo_svg_surface_get_document_unit}()},
\code{\link{cairo_svg_surface_set_document_unit}()},
\code{\link{cairo_svg_version_to_string}()}
}
\concept{cairo-svg-surface}
