% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-svg-enums.R
\docType{data}
\name{cairo_svg_unit_t}
\alias{cairo_svg_unit_t}
\title{cairo_svg_unit_t}
\format{
An object of class \code{list} of length 10.
}
\usage{
cairo_svg_unit_t
}
\description{
#cairo_svg_unit_t is used to describe the units valid for coordinates and
lengths in the SVG specification.
}
\details{
See also:
https://www.w3.org/TR/SVG/coords.html#Units
https://www.w3.org/TR/SVG/types.html#DataTypeLength
https://www.w3.org/TR/css-values-3/#lengths

Since: 1.16

\describe{
\item{CAIRO_SVG_UNIT_USER}{User unit, a value in the current coordinate system. If used in the root element for the initial coordinate systems it corresponds to pixels. (Since 1.16)}
\item{CAIRO_SVG_UNIT_EM}{The size of the element's font. (Since 1.16)}
\item{CAIRO_SVG_UNIT_EX}{The x-height of the element’s font. (Since 1.16)}
\item{CAIRO_SVG_UNIT_PX}{Pixels (1px = 1/96th of 1in). (Since 1.16)}
\item{CAIRO_SVG_UNIT_IN}{Inches (1in = 2.54cm = 96px). (Since 1.16)}
\item{CAIRO_SVG_UNIT_CM}{Centimeters (1cm = 96px/2.54). (Since 1.16)}
\item{CAIRO_SVG_UNIT_MM}{Millimeters (1mm = 1/10th of 1cm). (Since 1.16)}
\item{CAIRO_SVG_UNIT_PT}{Points (1pt = 1/72th of 1in). (Since 1.16)}
\item{CAIRO_SVG_UNIT_PC}{Picas (1pc = 1/6th of 1in). (Since 1.16)}
\item{CAIRO_SVG_UNIT_PERCENT}{Percent, a value that is some fraction of another reference value. (Since 1.16)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_version_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
