% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-svg-enums.R
\docType{data}
\name{cairo_svg_version_t}
\alias{cairo_svg_version_t}
\title{cairo_svg_version_t}
\format{
An object of class \code{list} of length 2.
}
\usage{
cairo_svg_version_t
}
\description{
#cairo_svg_version_t is used to describe the version number of the SVG
specification that a generated SVG file will conform to.
}
\details{
Since: 1.2

\describe{
\item{CAIRO_SVG_VERSION_1_1}{The version 1.1 of the SVG specification. (Since 1.2)}
\item{CAIRO_SVG_VERSION_1_2}{The version 1.2 of the SVG specification. (Since 1.2)}
}
}
\seealso{
Other enum: 
\code{\link{cairo_antialias_t}},
\code{\link{cairo_content_t}},
\code{\link{cairo_device_type_t}},
\code{\link{cairo_extend_t}},
\code{\link{cairo_fill_rule_t}},
\code{\link{cairo_filter_t}},
\code{\link{cairo_font_slant_t}},
\code{\link{cairo_font_type_t}},
\code{\link{cairo_font_weight_t}},
\code{\link{cairo_format_t}},
\code{\link{cairo_ft_synthesize_t}},
\code{\link{cairo_hint_metrics_t}},
\code{\link{cairo_hint_style_t}},
\code{\link{cairo_int_status_t}},
\code{\link{cairo_line_cap_t}},
\code{\link{cairo_line_join_t}},
\code{\link{cairo_operator_t}},
\code{\link{cairo_path_data_type_t}},
\code{\link{cairo_pattern_type_t}},
\code{\link{cairo_pdf_metadata_t}},
\code{\link{cairo_pdf_outline_flags_t}},
\code{\link{cairo_pdf_version_t}},
\code{\link{cairo_ps_level_t}},
\code{\link{cairo_region_overlap_t}},
\code{\link{cairo_status_t}},
\code{\link{cairo_subpixel_order_t}},
\code{\link{cairo_surface_observer_mode_t}},
\code{\link{cairo_surface_type_t}},
\code{\link{cairo_svg_unit_t}},
\code{\link{cairo_text_cluster_flags_t}},
\code{\link{enum_find}()},
\code{\link{enum_lookup}()}
}
\concept{enum}
\keyword{datasets}
