% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-svg-surface.R
\name{cairo_svg_version_to_string}
\alias{cairo_svg_version_to_string}
\title{cairo_svg_version_to_string}
\usage{
cairo_svg_version_to_string(version)
}
\arguments{
\item{version}{[\code{int}] a version id}
}
\value{
[\code{const char *}]
}
\description{
Get the string representation of the given version id. This function
will return \%NULL if version isn't valid. See cairo_svg_get_versions()
for a way to get the list of valid version ids.
}
\details{
Return value: the string associated to given version.

Since: 1.2

C function prototype: \code{const char * cairo_svg_version_to_string (cairo_svg_version_t version)}
}
\seealso{
Other cairo-svg-surface: 
\code{\link{cairo_svg_surface_create}()},
\code{\link{cairo_svg_surface_get_document_unit}()},
\code{\link{cairo_svg_surface_restrict_to_version}()},
\code{\link{cairo_svg_surface_set_document_unit}()}
}
\concept{cairo-svg-surface}
